// my-app/src/generated/models/Spaar_materialrequestlinesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Spaar_materialrequestlinesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Spaar_materialrequestlinesstatecode = keyof typeof Spaar_materialrequestlinesstatecode;
export const Spaar_materialrequestlinesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Spaar_materialrequestlinesstatuscode = keyof typeof Spaar_materialrequestlinesstatuscode;

export interface Spaar_materialrequestlinesBase {
  importsequencenumber?: string;
  overriddencreatedon?: string;
  spaar_materialid?: string;
  spaar_materialname?: string;
  "spaar_MaterialRequest@odata.bind"?: string;
  spaar_materialrequestline1: string;
  spaar_materialrequestlineid: string;
  spaar_productcode?: string;
  spaar_quantity?: string;
  spaar_unit?: string;
  statecode: Spaar_materialrequestlinesstatecode;
  statuscode?: Spaar_materialrequestlinesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Spaar_materialrequestlines extends Spaar_materialrequestlinesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  spaar_materialrequestname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
  spaar_materialrequest?: object;
  _spaar_materialrequest_value?: string;
}